refactor: decompose tui/app.rs into focused state modules (#433)
Extract four sub-modules from app.rs (1176 -> 720 LOC):
- scroll_controller.rs: ScrollController struct + impl (75 LOC)
- app_tools.rs: tool execution methods as impl App (106 LOC)
- app_input.rs: input delegation methods as impl App (120 LOC)
- focus_manager.rs: LayoutCache, PaneWrapper, focus caching (188 LOC)

Fields accessed by sibling impl blocks use pub(super) visibility.
All imports updated (debug_panel.rs, mod.rs re-exports).
